requirements.txt
langchain==0.0.267
langchain-openai==0.0.2
openai==1.3.5
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
python-dotenv==1.0.0
pypdf==3.15.1
streamlit==1.24.0 